+++ examples/simple.rs
@@ -68,9 +68,11 @@ impl InputContext for PlayerBox {
         // be triggered with any non-zero value.
         ctx.bind::<Move>()
             .to((Cardinal::wasd_keys(), GamepadStick::Left))
-            .with_modifier(DeadZone::default()) // Apply non-uniform normalization to ensure consistent speed, otherwise diagonal movement will be faster.
-            .with_modifier(DeltaLerp::default()) // Make movement smooth and independent of the framerate. To only make it framerate-independent, use `DeltaScale`.
-            .with_modifier(Scale::splat(DEFAULT_SPEED)); // Additionally multiply by a constant to achieve the desired speed.
+            .with_modifiers((
+                DeadZone::default(), // Apply non-uniform normalization to ensure consistent speed, otherwise diagonal movement will be faster.
+                DeltaLerp::default(), // Make movement smooth and independent of the framerate. To only make it framerate-independent, use `DeltaScale`.
+                Scale::splat(DEFAULT_SPEED), // Additionally multiply by a constant to achieve the desired speed.
+            ));
 
         ctx.bind::<Rotate>()
             .to((KeyCode::Space, GamepadButtonType::South));