+++ src/input_context/input_modifier.rs
@@ -7,7 +7,7 @@ pub mod negate;
 pub mod scale;
 pub mod swizzle_axis;
 
-use std::fmt::Debug;
+use std::{fmt::Debug, iter};
 
 use bevy::prelude::*;
 
@@ -20,8 +20,8 @@ use crate::action_value::ActionValue;
 /// or changing how input maps to axes.
 ///
 /// Can be applied both to inputs and actions.
-/// See [`ActionBind::with_modifier`](super::context_instance::ActionBind::with_modifier)
-/// and [`InputBindModCond::with_modifier`](super::input_bind::InputBindModCond::with_modifier).
+/// See [`ActionBind::with_modifiers`](super::context_instance::ActionBind::with_modifiers)
+/// and [`InputBindModCond::with_modifiers`](super::input_bind::InputBindModCond::with_modifiers).
 pub trait InputModifier: Sync + Send + Debug + 'static {
     /// Returns pre-processed value.
     ///
@@ -33,3 +33,35 @@ pub trait InputModifier: Sync + Send + Debug + 'static {
         value: ActionValue,
     ) -> ActionValue;
 }
+
+/// Represents collection of bindings that could be passed into
+/// [`ActionBind::with_modifiers`](super::context_instance::ActionBind::with_modifiers)
+/// and [`InputBindModCond::with_modifiers`](super::input_bind::InputBindModCond::with_modifiers).
+pub trait InputModifiers {
+    /// Returns an iterator over modifiers.
+    fn iter_modifiers(self) -> impl Iterator<Item = Box<dyn InputModifier>>;
+}
+
+impl<I: InputModifier> InputModifiers for I {
+    fn iter_modifiers(self) -> impl Iterator<Item = Box<dyn InputModifier>> {
+        iter::once(Box::new(self) as Box<dyn InputModifier>)
+    }
+}
+
+macro_rules! impl_tuple_modifiers {
+    ($($name:ident),+) => {
+        impl<$($name),+> InputModifiers for ($($name,)+)
+        where
+            $($name: InputModifier),+
+        {
+            #[allow(non_snake_case)]
+            fn iter_modifiers(self) -> impl Iterator<Item = Box<dyn InputModifier>> {
+                let ($($name,)+) = self;
+                std::iter::empty()
+                    $(.chain(iter::once(Box::new($name) as Box<dyn InputModifier>)))+
+            }
+        }
+    };
+}
+
+bevy::utils::all_tuples!(impl_tuple_modifiers, 1, 15, I);