+++ CHANGELOG.md
@@ -19,7 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Rename `Input::with_modifiers` and `Input::without_modifiers` into `Input::with_mod_keys` and `Input::without_mod_keys`, respectively.
 - `Input::with_mod_keys` now a trait method which implemented for any type that can be converted into an input to ergonomically assign keyboard modifiers.
 - `Input::without_mod_keys` no longer `const`.
-- `InputBind::with_modifier` and `InputBind::with_condition` now trait methods which implemented for any type that can be converted into a binding to ergonomically assign modifiers and conditions.
+- Rename all `with_modifier` and `with_condition` methods into `with_modifiers` and `with_conditions` (plural) because now they accept tuples.
+- `InputBind::with_modifiers` and `InputBind::with_conditions` now trait methods which implemented for any type that can be converted into a binding to ergonomically assign modifiers and conditions.
 - Move `ActionState`, `ActionData` and `ActionsData` to `context_instance` module.
 - Move `InputBind` to newly created `input_context::input_bind` module.